requirements.txt
numpy~=1.24.1
matplotlib~=3.6.2
pwntools~=4.9.0